fix+cleanup: properly report errors
* factor out logger so it is not dependent upon electron
* properly send errors to a dialog box
* fix: export dialog should close when you hit 'ok'
