+++ index.js
@@ -9,13 +9,13 @@ const debug = require('electron-debug')
 const mkdirp = require('mkdirp')
 const series = require('run-series')
 const styles = require('mapeo-styles')
-const logger = require('electron-timber')
 const { fork } = require('child_process')
+const middleware = require('hoist')
 
 const app = electron.app
 const BrowserWindow = electron.BrowserWindow
 
-const middleware = require('./src/middleware')
+const logger = require('./src/logger')
 const miscellaneousIpc = require('./src/main/ipc')
 const createMenu = require('./src/main/menu')
 const windowStateKeeper = require('./src/main/window-state')
@@ -83,6 +83,10 @@ function openWindow () {
   // TODO: get a socket name that isn't open ..
   var _socketName = 'mapeo1'
   logger.log('got socket', _socketName)
+  ipc.on('error', function (err) {
+    logger.error(err)
+    electron.dialog.showErrorBox('Error', err)
+  })
   ipc.connect(_socketName)
 
   if (!win) {
@@ -215,7 +219,7 @@ function createWindow (socketName) {
     icon: path.resolve(__dirname, 'static', 'mapeo_256x256.png'),
     webPreferences: {
       nodeIntegration: true,
-      preload: path.resolve(__dirname, 'src', 'middleware', 'client-preload.js')
+      preload: path.resolve(__dirname, 'src', 'windows', 'index-preload.js')
     }
   })
   mainWindowState.manage(win)
@@ -245,7 +249,7 @@ function createBgWindow (socketName) {
     }
   })
   console.log('loading bg window')
-  var BG = 'file://' + path.join(__dirname, '/background.html')
+  var BG = 'file://' + path.join(__dirname, './src/windows/background.html')
   win.loadURL(BG)
   win.webContents.on('did-finish-load', () => {
     if (argv.debug) bg.webContents.openDevTools()
@@ -261,22 +265,22 @@ function createBgWindow (socketName) {
 }
 
 function createSplashWindow () {
-  var SPLASH = 'file://' + path.join(__dirname, './splash.html')
+  var SPLASH = 'file://' + path.join(__dirname, './src/windows/splash.html')
   var splash = new BrowserWindow({
-      width: 450,
-      height: 410,
-      center: true,
-      transparent: true,
-      resizable: false,
-      frame: false
-    })
+    width: 450,
+    height: 410,
+    center: true,
+    transparent: true,
+    resizable: false,
+    frame: false
+  })
   splash.loadURL(SPLASH)
   return splash
 }
 
 function createBackgroundProcess (socketName) {
   console.log('creating background process')
-  serverProcess = fork(path.join(__dirname, 'src', 'middleware', 'background.js'), [
+  serverProcess = fork(path.join(__dirname, 'src', 'windows', 'background.js'), [
     '--subprocess',
     app.getVersion(),
     socketName
