+++ package.json
@@ -89,6 +89,7 @@
     "electron-timber": "^0.5.1",
     "flat": "^5.0.0",
     "fs-write-stream-atomic": "^1.0.10",
+    "hoist": "github:okdistribute/hoist",
     "html-react-parser": "^0.9.1",
     "id-mapeo": "^3.1.6",
     "insert-css": "^1.0.0",
@@ -99,7 +100,6 @@
     "mapeo-styles": "^3.0.0",
     "minimist": "^1.2.0",
     "mkdirp": "^0.5.1",
-    "node-ipc": "^9.1.1",
     "once": "^1.4.0",
     "osm-p2p": "^5.1.0",
     "osm-p2p-server": "^5.0.0",
