+++ src/main/ipc.js
@@ -1,7 +1,7 @@
 var path = require('path')
 var { dialog, app, ipcMain } = require('electron')
-var logger = require('electron-timber')
 
+var logger = require('../logger')
 var userConfig = require('./user-config')
 var i18n = require('./i18n')
 
@@ -9,6 +9,8 @@ var i18n = require('./i18n')
  * Miscellaneous ipc calls that don't hit mapeo-core
  */
 module.exports = function (win) {
+  var ipc = ipcMain
+
   function ipcSend (...args) {
     try {
       win.webContents.send.apply(win.webContents, args)
@@ -17,21 +19,21 @@ module.exports = function (win) {
     }
   }
 
-  ipcMain.on('get-user-data', function (event, type) {
+  ipc.on('get-user-data', function (event, type) {
     var data = userConfig.getSettings(type)
     if (!data) console.warn('unhandled event', type)
     event.returnValue = data
   })
 
-  ipcMain.on('error', function (ev, message) {
+  ipc.on('error', function (ev, message) {
     ipcSend('error', message)
   })
 
-  ipcMain.on('set-locale', function (ev, lang) {
+  ipc.on('set-locale', function (ev, lang) {
     app.translations = i18n.setLocale(lang)
   })
 
-  ipcMain.on('import-example-presets', function (ev) {
+  ipc.on('import-example-presets', function (ev) {
     var filename = path.join(
       __dirname,
       '..',
@@ -45,14 +47,14 @@ module.exports = function (win) {
     })
   })
 
-  ipcMain.on('import-settings', function (ev, filename) {
+  ipc.on('import-settings', function (ev, filename) {
     userConfig.importSettings(win, filename, function (err) {
       if (err) return logger.error(err)
       logger.log('Example presets imported from ' + filename)
     })
   })
 
-  ipcMain.on('save-file', function () {
+  ipc.on('save-file', function () {
     var metadata = userConfig.getSettings('metadata')
     var ext = metadata ? metadata.dataset_id : 'mapeodata'
     dialog.showSaveDialog(
@@ -75,7 +77,7 @@ module.exports = function (win) {
     }
   })
 
-  ipcMain.on('open-file', function () {
+  ipc.on('open-file', function () {
     var metadata = userConfig.getSettings('metadata')
     var ext = metadata ? metadata.dataset_id : 'mapeodata'
     dialog.showOpenDialog(
@@ -101,15 +103,15 @@ module.exports = function (win) {
     }
   })
 
-  ipcMain.on('zoom-to-latlon-request', function (_, lon, lat) {
+  ipc.on('zoom-to-latlon-request', function (_, lon, lat) {
     ipcSend('zoom-to-latlon-response', [lon, lat])
   })
 
-  ipcMain.on('force-refresh-window', function () {
+  ipc.on('force-refresh-window', function () {
     ipcSend('force-refresh-window')
   })
 
-  ipcMain.on('refresh-window', function () {
+  ipc.on('refresh-window', function () {
     ipcSend('refresh-window')
   })
 }
