+++ src/main/tile-server.js
@@ -1,7 +1,8 @@
 var http = require('http')
 var ecstatic = require('ecstatic')
 var series = require('run-series')
-var logger = console
+
+var logger = require('../logger')
 
 module.exports = function (userDataPath) {
   var guesses = ['png', 'jpg', 'jpeg']
