--- src/middleware/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-  Client: require('./client-ipc'),
-  findOpenSocket: require('./find-open-socket')
-}
