+++ src/renderer/components/MapFilter/MapFilter.js
@@ -3,12 +3,12 @@ import { makeStyles } from '@material-ui/core/styles'
 import { MapView, ReportView, MediaView } from 'react-mapfilter'
 import { ipcRenderer } from 'electron'
 import debounce from 'lodash/debounce'
-import logger from 'electron-timber'
 import { defineMessages, FormattedMessage } from 'react-intl'
 import mapboxgl from 'mapbox-gl'
 import ErrorBoundary from 'react-error-boundary'
 import { Typography } from '@material-ui/core'
 
+import logger from '../../../logger'
 import Toolbar from './Toolbar'
 import FilterPanel from './FilterPanel'
 import Loading from './Loading'
