+++ src/windows/background.html
@@ -4,7 +4,7 @@
   </head>
   <body>
   <script>
-      require('./src/middleware/background.js')
+      require('./background.js')
   </script>
   </body>
 </html>
