+++ src/windows/index-preload.js
@@ -1,5 +1,5 @@
 const { ipcRenderer } = require('electron')
-const middleware = require('./')
+const middleware = require('hoist')
 
 window.middlewareClient = new middleware.Client()
 ipcRenderer.on('set-socket', (event, { name }) => {
