+++ src/windows/splash.html
@@ -2,7 +2,7 @@
 <html>
   <head>
     <meta charset="UTF-8">
-    <link rel='stylesheet' href='static/css/base.css'>
+    <link rel='stylesheet' href='../../static/css/base.css'>
     <meta name="viewport" content="minimum-scale=1, initial-scale=1, width=device-width, shrink-to-fit=no"/>
     <meta name='apple-mobile-web-app-capable' content='yes'/>
     <meta name='apple-mobile-web-app-status-bar-style' content='black-translucent'/>
@@ -10,7 +10,7 @@
   <body>
   <div id="welcome" class='mapeo-container'>
     <div class="mapeo-splash">
-      <img src="static/mapeo-white.svg" class="mapeo-splash-logo" />
+      <img src="../../static/mapeo-white.svg" class="mapeo-splash-logo" />
       <div class="mapeo-splash-loader" aria-label="Loading..."></div>
     </div>
   </div>
